LPS-132649 Use SPI to load up ModuleFrameworkImpl to avoid the reflection bridge.
