+++ modules/core/portal-bootstrap/src/main/resources/META-INF/services/com.liferay.portal.module.framework.ModuleFramework
@@ -1,0 +1,1 @@
+com.liferay.portal.bootstrap.ModuleFrameworkImpl
