+++ portal-impl/src/com/liferay/portal/tools/DBUpgrader.java
@@ -36,7 +36,7 @@ import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.ReleaseInfo;
 import com.liferay.portal.kernel.util.Time;
 import com.liferay.portal.kernel.version.Version;
-import com.liferay.portal.module.framework.ModuleFrameworkUtilAdapter;
+import com.liferay.portal.module.framework.ModuleFrameworkUtil;
 import com.liferay.portal.transaction.TransactionsUtil;
 import com.liferay.portal.upgrade.PortalUpgradeProcess;
 import com.liferay.portal.util.InitUtil;
@@ -294,7 +294,7 @@ public class DBUpgrader {
 			InitUtil.registerContext();
 		}
 		else {
-			ModuleFrameworkUtilAdapter.registerContext(applicationContext);
+			ModuleFrameworkUtil.registerContext(applicationContext);
 		}
 
 		_registerModuleServiceLifecycle("portal.initialized");
